# docs/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DeNoise application API",
    "description": "Authenticated operator API (docs/06-api-contract.md). Session cookie for the UI or \u0060Authorization: Bearer ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E\u0060 for API clients.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/heartbeats": {
      "get": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "ListHeartbeats",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "integration",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartbeatSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "CreateHeartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/preview": {
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "PreviewHeartbeatSchedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/export": {
      "get": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "ExportHeartbeats",
        "parameters": [
          {
            "name": "team",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/import": {
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "ImportHeartbeats",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatImportResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/{id}": {
      "get": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "GetHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "put": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "UpdateHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "DeleteHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/{id}/pause": {
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "PauseHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/{id}/resume": {
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "ResumeHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/heartbeats/{id}/rotate-token": {
      "post": {
        "tags": [
          "Heartbeats"
        ],
        "operationId": "RotateHeartbeatToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingUrlResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HeartbeatCreatedResponse": {
        "required": [
          "heartbeat",
          "pingUrl"
        ],
        "type": "object",
        "properties": {
          "heartbeat": {
            "$ref": "#/components/schemas/HeartbeatSummary"
          },
          "pingUrl": {
            "type": "string"
          }
        }
      },
...
      "HeartbeatRequest": {
        "required": [
          "name",
          "description",
          "owningTeamId",
          "assigneeId",
          "schedule",
          "grace",
          "severityOnMiss",
          "routingPolicyId",
          "bindsToIntegrationId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "owningTeamId": {
            "type": "string",
            "format": "uuid"
          },
          "assigneeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "schedule": {
            "$ref": "#/components/schemas/HeartbeatScheduleRequest"
          },
          "grace": {
            "type": "string"
          },
          "severityOnMiss": {
            "type": "string"
          },
          "routingPolicyId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "bindsToIntegrationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "recoverySuccessesRequired": {
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "autoPauseDuringMaintenance": {
            "type": "boolean",
            "default": true
          },
          "accessScope": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "HeartbeatSummary": {
        "required": [
          "id",
          "name",
          "description",
          "accessScope",
          "owningTeam",
          "assigneeId",
          "schedule",
          "grace",
          "severityOnMiss",
          "routingPolicyId",
          "bindsToIntegrationId",
          "boundIntegrationName",
          "recoverySuccessesRequired",
          "autoPauseDuringMaintenance",
          "state",
          "expectedNext",
          "lastPingAt",
          "lastPingIp",
          "lastRunDuration",
          "pausedBy",
          "pausedAt",
          "pauseReason",
          "pausedByMaintenance",
          "missEpisodeId",
          "keyId",
          "tokenRotatedAt",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "accessScope": {
            "type": "string"
          },
          "owningTeam": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TeamRef"
              }
            ]
          },
          "assigneeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "schedule": {
            "$ref": "#/components/schemas/HeartbeatScheduleDto"
          },
          "grace": {
            "type": "string"
          },
          "severityOnMiss": {
            "type": "string"
          },
          "routingPolicyId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "bindsToIntegrationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "boundIntegrationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "recoverySuccessesRequired": {
            "type": "integer",
            "format": "int32"
          },
          "autoPauseDuringMaintenance": {
            "type": "boolean"
          },
          "state": {
            "type": "string"
          },
          "expectedNext": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastPingAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastPingIp": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastRunDuration": {
            "type": [
              "null",
              "string"
            ]
          },
          "pausedBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "pausedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "pauseReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "pausedByMaintenance": {
            "type": "boolean"
          },
          "missEpisodeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "keyId": {
            "type": "string"
          },
          "tokenRotatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SchedulePreviewRequest": {
        "required": [
          "schedule"
        ],
        "type": "object",
        "properties": {
          "schedule": {
            "$ref": "#/components/schemas/HeartbeatScheduleRequest"
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "default": 5
          }
        }
      },
      "SchedulePreviewResponse": {
        "required": [
          "nextRuns",
          "description"
        ],
        "type": "object",
        "properties": {
          "nextRuns": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Server-side session cookie set by POST /auth/login.",
        "name": "denoise_session",
        "in": "cookie"
      },
      "pat": {
        "type": "http",
        "description": "Personal access token ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E.",
        "scheme": "bearer"
      }
    }
  },
...
}